Move unsupported types codec bindings to its own file
